chunk dataframes so easy to split for parallel processing

chunk dataframes so easy to split for parallel processing

chunk_df(.gdf, ..., .nchunks = NULL)

Arguments

.gdf
(grouped) data frame

...
grouping variables, either a character vector or NSE-style column names

.nchunks
set number of chunks

Examples

library(dplyr)
Theoph %>% chunk_df()

#>     Subject   Wt Dose  Time  conc chunk__
#> 1         1 79.6 4.02  0.00  0.74       1
#> 2         1 79.6 4.02  0.25  2.84       1
#> 3         1 79.6 4.02  0.57  6.57       1
#> 4         1 79.6 4.02  1.12 10.50       1
#> 5         1 79.6 4.02  2.02  9.66       1
#> 6         1 79.6 4.02  3.82  8.58       1
#> 7         1 79.6 4.02  5.10  8.36       1
#> 8         1 79.6 4.02  7.03  7.47       1
#> 9         1 79.6 4.02  9.05  6.89       1
#> 10        1 79.6 4.02 12.12  5.94       1
#> 11        1 79.6 4.02 24.37  3.28       1
#> 12        2 72.4 4.40  0.00  0.00       1
#> 13        2 72.4 4.40  0.27  1.72       1
#> 14        2 72.4 4.40  0.52  7.91       1
#> 15        2 72.4 4.40  1.00  8.31       1
#> 16        2 72.4 4.40  1.92  8.33       1
#> 17        2 72.4 4.40  3.50  6.85       1
#> 18        2 72.4 4.40  5.02  6.08       2
#> 19        2 72.4 4.40  7.03  5.40       2
#> 20        2 72.4 4.40  9.00  4.55       2
#> 21        2 72.4 4.40 12.00  3.01       2
#> 22        2 72.4 4.40 24.30  0.90       2
#> 23        3 70.5 4.53  0.00  0.00       2
#> 24        3 70.5 4.53  0.27  4.40       2
#> 25        3 70.5 4.53  0.58  6.90       2
#> 26        3 70.5 4.53  1.02  8.20       2
#> 27        3 70.5 4.53  2.02  7.80       2
#> 28        3 70.5 4.53  3.62  7.50       2
#> 29        3 70.5 4.53  5.08  6.20       2
#> 30        3 70.5 4.53  7.07  5.30       2
#> 31        3 70.5 4.53  9.00  4.90       2
#> 32        3 70.5 4.53 12.15  3.70       2
#> 33        3 70.5 4.53 24.17  1.05       2
#> 34        4 72.7 4.40  0.00  0.00       2
#> 35        4 72.7 4.40  0.35  1.89       3
#> 36        4 72.7 4.40  0.60  4.60       3
#> 37        4 72.7 4.40  1.07  8.60       3
#> 38        4 72.7 4.40  2.13  8.38       3
#> 39        4 72.7 4.40  3.50  7.54       3
#> 40        4 72.7 4.40  5.02  6.88       3
#> 41        4 72.7 4.40  7.02  5.78       3
#> 42        4 72.7 4.40  9.02  5.33       3
#> 43        4 72.7 4.40 11.98  4.19       3
#> 44        4 72.7 4.40 24.65  1.15       3
#> 45        5 54.6 5.86  0.00  0.00       3
#> 46        5 54.6 5.86  0.30  2.02       3
#> 47        5 54.6 5.86  0.52  5.63       3
#> 48        5 54.6 5.86  1.00 11.40       3
#> 49        5 54.6 5.86  2.02  9.33       3
#> 50        5 54.6 5.86  3.50  8.74       3
#> 51        5 54.6 5.86  5.02  7.56       3
#> 52        5 54.6 5.86  7.02  7.09       4
#> 53        5 54.6 5.86  9.10  5.90       4
#> 54        5 54.6 5.86 12.00  4.37       4
#> 55        5 54.6 5.86 24.35  1.57       4
#> 56        6 80.0 4.00  0.00  0.00       4
#> 57        6 80.0 4.00  0.27  1.29       4
#> 58        6 80.0 4.00  0.58  3.08       4
#> 59        6 80.0 4.00  1.15  6.44       4
#> 60        6 80.0 4.00  2.03  6.32       4
#> 61        6 80.0 4.00  3.57  5.53       4
#> 62        6 80.0 4.00  5.00  4.94       4
#> 63        6 80.0 4.00  7.00  4.02       4
#> 64        6 80.0 4.00  9.22  3.46       4
#> 65        6 80.0 4.00 12.10  2.78       4
#> 66        6 80.0 4.00 23.85  0.92       4
#> 67        7 64.6 4.95  0.00  0.15       4
#> 68        7 64.6 4.95  0.25  0.85       4
#> 69        7 64.6 4.95  0.50  2.35       5
#> 70        7 64.6 4.95  1.02  5.02       5
#> 71        7 64.6 4.95  2.02  6.58       5
#> 72        7 64.6 4.95  3.48  7.09       5
#> 73        7 64.6 4.95  5.00  6.66       5
#> 74        7 64.6 4.95  6.98  5.25       5
#> 75        7 64.6 4.95  9.00  4.39       5
#> 76        7 64.6 4.95 12.05  3.53       5
#> 77        7 64.6 4.95 24.22  1.15       5
#> 78        8 70.5 4.53  0.00  0.00       5
#> 79        8 70.5 4.53  0.25  3.05       5
#> 80        8 70.5 4.53  0.52  3.05       5
#> 81        8 70.5 4.53  0.98  7.31       5
#> 82        8 70.5 4.53  2.02  7.56       5
#> 83        8 70.5 4.53  3.53  6.59       5
#> 84        8 70.5 4.53  5.05  5.88       5
#> 85        8 70.5 4.53  7.15  4.73       6
#> 86        8 70.5 4.53  9.07  4.57       6
#> 87        8 70.5 4.53 12.10  3.00       6
#> 88        8 70.5 4.53 24.12  1.25       6
#> 89        9 86.4 3.10  0.00  0.00       6
#> 90        9 86.4 3.10  0.30  7.37       6
#> 91        9 86.4 3.10  0.63  9.03       6
#> 92        9 86.4 3.10  1.05  7.14       6
#> 93        9 86.4 3.10  2.02  6.33       6
#> 94        9 86.4 3.10  3.53  5.66       6
#> 95        9 86.4 3.10  5.02  5.67       6
#> 96        9 86.4 3.10  7.17  4.24       6
#> 97        9 86.4 3.10  8.80  4.11       6
#> 98        9 86.4 3.10 11.60  3.16       6
#> 99        9 86.4 3.10 24.43  1.12       6
#> 100      10 58.2 5.50  0.00  0.24       6
#> 101      10 58.2 5.50  0.37  2.89       7
#> 102      10 58.2 5.50  0.77  5.22       7
#> 103      10 58.2 5.50  1.02  6.41       7
#> 104      10 58.2 5.50  2.05  7.83       7
#> 105      10 58.2 5.50  3.55 10.21       7
#> 106      10 58.2 5.50  5.05  9.18       7
#> 107      10 58.2 5.50  7.08  8.02       7
#> 108      10 58.2 5.50  9.38  7.14       7
#> 109      10 58.2 5.50 12.10  5.68       7
#> 110      10 58.2 5.50 23.70  2.42       7
#> 111      11 65.0 4.92  0.00  0.00       7
#> 112      11 65.0 4.92  0.25  4.86       7
#> 113      11 65.0 4.92  0.50  7.24       7
#> 114      11 65.0 4.92  0.98  8.00       7
#> 115      11 65.0 4.92  1.98  6.81       7
#> 116      11 65.0 4.92  3.60  5.87       7
#> 117      11 65.0 4.92  5.02  5.22       8
#> 118      11 65.0 4.92  7.03  4.45       8
#> 119      11 65.0 4.92  9.03  3.62       8
#> 120      11 65.0 4.92 12.12  2.69       8
#> 121      11 65.0 4.92 24.08  0.86       8
#> 122      12 60.5 5.30  0.00  0.00       8
#> 123      12 60.5 5.30  0.25  1.25       8
#> 124      12 60.5 5.30  0.50  3.96       8
#> 125      12 60.5 5.30  1.00  7.82       8
#> 126      12 60.5 5.30  2.00  9.72       8
#> 127      12 60.5 5.30  3.52  9.75       8
#> 128      12 60.5 5.30  5.07  8.57       8
#> 129      12 60.5 5.30  7.07  6.59       8
#> 130      12 60.5 5.30  9.03  6.11       8
#> 131      12 60.5 5.30 12.05  4.57       8
#> 132      12 60.5 5.30 24.15  1.17       8

Theoph %>% group_by(Subject) %>% chunk_df()

#> # A tibble: 132 x 6
#> # Groups:   Subject [12]
#>    Subject    Wt  Dose   Time  conc chunk__
#>    <ord>   <dbl> <dbl>  <dbl> <dbl>   <int>
#>  1 1        79.6  4.02  0      0.74       1
#>  2 1        79.6  4.02  0.25   2.84       1
#>  3 1        79.6  4.02  0.570  6.57       1
#>  4 1        79.6  4.02  1.12  10.5        1
#>  5 1        79.6  4.02  2.02   9.66       1
#>  6 1        79.6  4.02  3.82   8.58       1
#>  7 1        79.6  4.02  5.1    8.36       1
#>  8 1        79.6  4.02  7.03   7.47       1
#>  9 1        79.6  4.02  9.05   6.89       1
#> 10 1        79.6  4.02 12.1    5.94       1
#> # ... with 122 more rows

Theoph %>% chunk_df(Subject)

#> # A tibble: 132 x 6
#> # Groups:   Subject [12]
#>    Subject    Wt  Dose   Time  conc chunk__
#>    <ord>   <dbl> <dbl>  <dbl> <dbl>   <int>
#>  1 1        79.6  4.02  0      0.74       1
#>  2 1        79.6  4.02  0.25   2.84       1
#>  3 1        79.6  4.02  0.570  6.57       1
#>  4 1        79.6  4.02  1.12  10.5        1
#>  5 1        79.6  4.02  2.02   9.66       1
#>  6 1        79.6  4.02  3.82   8.58       1
#>  7 1        79.6  4.02  5.1    8.36       1
#>  8 1        79.6  4.02  7.03   7.47       1
#>  9 1        79.6  4.02  9.05   6.89       1
#> 10 1        79.6  4.02 12.1    5.94       1
#> # ... with 122 more rows

Theoph %>% chunk_df(c("Subject"))

#> # A tibble: 132 x 6
#> # Groups:   Subject [12]
#>    Subject    Wt  Dose   Time  conc chunk__
#>    <ord>   <dbl> <dbl>  <dbl> <dbl>   <int>
#>  1 1        79.6  4.02  0      0.74       1
#>  2 1        79.6  4.02  0.25   2.84       1
#>  3 1        79.6  4.02  0.570  6.57       1
#>  4 1        79.6  4.02  1.12  10.5        1
#>  5 1        79.6  4.02  2.02   9.66       1
#>  6 1        79.6  4.02  3.82   8.58       1
#>  7 1        79.6  4.02  5.1    8.36       1
#>  8 1        79.6  4.02  7.03   7.47       1
#>  9 1        79.6  4.02  9.05   6.89       1
#> 10 1        79.6  4.02 12.1    5.94       1
#> # ... with 122 more rows

Theoph %>% chunk_df(Subject, .nchunks = 3)

#> # A tibble: 132 x 6
#> # Groups:   Subject [12]
#>    Subject    Wt  Dose   Time  conc chunk__
#>    <ord>   <dbl> <dbl>  <dbl> <dbl>   <int>
#>  1 1        79.6  4.02  0      0.74       1
#>  2 1        79.6  4.02  0.25   2.84       1
#>  3 1        79.6  4.02  0.570  6.57       1
#>  4 1        79.6  4.02  1.12  10.5        1
#>  5 1        79.6  4.02  2.02   9.66       1
#>  6 1        79.6  4.02  3.82   8.58       1
#>  7 1        79.6  4.02  5.1    8.36       1
#>  8 1        79.6  4.02  7.03   7.47       1
#>  9 1        79.6  4.02  9.05   6.89       1
#> 10 1        79.6  4.02 12.1    5.94       1
#> # ... with 122 more rows