find all unique non-numeric values

find all unique non-numeric values

unique_non_numerics(x, na.rm = TRUE)

Arguments

x
vector to check on

na.rm
remove existing na values before checking

Details

This function is especially useful for figuring out what non-numeric unique values are in in a column that should be numeric so one can easily replace them with another flag. This function can work well with replace_char_flags instead of using nested ifelse statements

See also

replace_values: to use to replace non-numeric values in a dataframe.

Examples

dv <- c(1, 2, 4, "88 (excluded)", "bql", "*")
unique_non_numerics(dv)

#> [1] "88 (excluded)" "bql"           "*"            

df <- tibble::data_frame(ID = 1:3, DV = c("BQL", 0.5, 9))
unique_non_numerics(df$DV)

#> [1] "BQL"

#using dplyr
library(dplyr)
df %>% filter(!(DV %in% unique_non_numerics(DV)))

#> # A tibble: 2 x 2
#>      ID    DV
#>   <int> <chr>
#> 1     2   0.5
#> 2     3     9